+++ src/components/App/App.css
@@ -1,3 +1,4 @@
 .App {
   text-align: center;
 }
+
